extract modal helpers into a helper file
